+++ ADCDACPi/ABE_ADCDACPi.h
@@ -50,7 +50,7 @@ void close_dac();
 * When in differential mode setting channel to 2 will make IN1 = IN- and IN2 = IN+
 * @returns between 0V and the reference voltage
 */
-double read_adc_voltage(int channel, int mode);
+double read_adc_voltage(uint8_t channel, uint8_t mode);
 
 /**
 * Read the raw value from the ADC
@@ -60,7 +60,7 @@ double read_adc_voltage(int channel, int mode);
 * When in differential mode setting channel to 2 will make IN1 = IN- and IN2 = IN+
 * @returns 12 bit value between 0 and 4096
 */
-int read_adc_raw(int channel, int mode);
+uint16_t read_adc_raw(uint8_t channel, uint8_t mode);
 
 /**
 * Set the reference voltage for the adc
@@ -73,20 +73,20 @@ void set_adc_refvoltage(double ref);
 * @param voltage - between 0 and 2.048 when gain is set to 1,  0 and 3.3 when gain is set to 2
 * @param channel - 1 or 2
 */
-void set_dac_voltage(double voltage, int channel);
+void set_dac_voltage(double voltage, uint8_t channel);
 
 /**
 * Set the raw value from the selected channel on the DAC
 * @param raw - between 0 and 4095
 * @param channel - 1 or 2
 */
-void set_dac_raw(uint16_t raw, int channel);
+void set_dac_raw(uint16_t raw, uint8_t channel);
 
 /**
 * Set the DAC gain
 * @param gain - 1 or 2 - The output voltage will be between 0 and 2.048V when gain is set to 1,  0 and 3.3V when gain is set to 2	
 */
-void set_dac_gain(int gain);
+void set_dac_gain(uint8_t gain);
 
 private:
 // private variables
@@ -100,7 +100,7 @@ int adc; // adc object
 int dac; // dac object
 
 double adcrefvoltage; // reference voltage for the ADC chip.
-int dacgain; // gain setting for the DAC chip.
+uint8_t dacgain; // gain setting for the DAC chip.
 double dacvoltage; // maximum voltage for the DAC output
 
 };
